Adding schema validation after loading
